fix(carteira): omie_codigo_vendedor int → bigint (códigos Omie estouram int4)
Códigos reais do Omie (ex.: 8689670842 ≈ 8,6 bi) excedem int4 (~2,1 bi).
Tabelas já criadas no banco precisam de ALTER (entregue no bloco de seed).

+++ supabase/migrations/20260524120000_carteira_omie_fase1.sql
@@ -5,7 +5,7 @@
 CREATE TABLE IF NOT EXISTS public.omie_vendedor_map (
   id uuid PRIMARY KEY DEFAULT gen_random_uuid(),
   omie_account text NOT NULL,
-  omie_codigo_vendedor int NOT NULL,
+  omie_codigo_vendedor bigint NOT NULL,
   user_id uuid NOT NULL REFERENCES auth.users(id) ON DELETE CASCADE,
   nome text,
   created_at timestamptz NOT NULL DEFAULT now(),
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS public.carteira_assignments (
   owner_user_id uuid NOT NULL REFERENCES auth.users(id) ON DELETE CASCADE,
   source text NOT NULL CHECK (source IN ('omie','hunter_orphan')),
   omie_account text,
-  omie_codigo_vendedor int,
+  omie_codigo_vendedor bigint,
   eligible boolean NOT NULL DEFAULT true,
   valid_from timestamptz NOT NULL DEFAULT now(),
   updated_at timestamptz NOT NULL DEFAULT now(),
